Move veracity components to separate package
